render: address nits from Opus review of cull-effectiveness diagnostic
Fix three documentation-accuracy issues identified by Opus recheck:

- system_voxel_to_trixel.hpp: replace "sync-free" with "no explicit fence"
  and correct "discarded" to "contributing a 1/N bias" — the first frame's
  zero sample is not discarded; it pulls the average down by ~1/N over an
  N-frame run.
- world.cpp (getVoxelCullStats Lua binding): promote uint64_t/uint32_t
  integer division to double so the Lua surface matches the precision of
  the C++ profile-report writer.
- docs/perf/README.md: same sync-semantics clarification as the source
  comment.

+++ engine/world/src/world.cpp
@@ -160,8 +160,10 @@ void World::setupLuaBindings(const std::vector<LuaBindingRegistration> &bindings
         out["visible"] = timing.visibleVoxelCount_;
         out["total"] = timing.totalVoxelCount_;
         out["samples"] = acc.sampleCount_;
-        out["avgVisible"] = acc.sampleCount_ > 0 ? acc.visibleSum_ / acc.sampleCount_ : 0;
-        out["avgTotal"] = acc.sampleCount_ > 0 ? acc.totalSum_ / acc.sampleCount_ : 0;
+        out["avgVisible"] =
+            acc.sampleCount_ > 0 ? static_cast<double>(acc.visibleSum_) / acc.sampleCount_ : 0.0;
+        out["avgTotal"] =
+            acc.sampleCount_ > 0 ? static_cast<double>(acc.totalSum_) / acc.sampleCount_ : 0.0;
         out["maxVisible"] = acc.maxVisible_;
         out["maxTotal"] = acc.maxTotal_;
         return out;
